+++ .gitignore
@@ -56,7 +56,6 @@ typings/
 
 # firebase  cache
 .firebase
-.runtimeconfig.json
 
 # TypeScript cache
 *.tsbuildinfo
