+++ nuxt.config.js
@@ -1,7 +1,4 @@
-// const functions = require("firebase-functions");
-// const firebaseConfigProject = (functions.config().projects || {}).zeboard || {};
-// console.log('nuxtModules=>nuxt.js: ',firebaseConfigProject,' full CONFIGs:', functions.config());
-import modules from "./nuxtModules.js";
+
 export default {
     mode: "spa",
     dev: "develop",
@@ -18,9 +15,11 @@ export default {
 
     plugins: [],
 
-    modules,
-   
-    buildModules: [],
+    modules: [
+        '~/modules/firebaseModule.js',
+    ],
+    buildModules: [
+    ],
 
-    target: "static",
+    target: 'static'
 };
